refactor(MockFormatter): update shared types and add returnValue formatter
- Updated type imports to use shared types directory
- Added and exported new type aliases for MockFormatter constructor and function
- Removed redundant file header comments and unused type exports
- Fixed test imports to align with new shared type locations
- Added `returnValue` formatter that always outputs a fixed string
