+++ packages/@agla-utils/ag-logger/src/plugins/formatter/__tests__/MockFormatter.spec.ts
@@ -9,7 +9,7 @@ import { describe, expect, it } from 'vitest';
 // types
 import { AG_LOGLEVEL } from '../../../../shared/types';
 import type { AgFormatFunction, AgLogMessage } from '../../../../shared/types';
-import type { AgFormatRoutine } from '../../../internal/types/AgMockConstructor.class';
+import type { AgFormatRoutine } from '../../../../shared/types/AgMockConstructor.class';
 
 // target
 import { createMockFormatter, MockFormatter } from '../MockFormatter';
